# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project_Alan API documentation example",
        "contact": {
            "email": "aldaberdiev99@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
        "/actions": {
            "get": {
                "tags": [
                    "Actions"
                ],
                "summary": "Display all actions",
                "operationId": "actionsAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ActionShowRequest"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Actions"
                ],
                "summary": "Store actions data",
                "description": "Store actions data",
                "operationId": "actionStore",
                "requestBody": {
                    "description": "Pass action credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "price",
                                    "title"
                                ],
                                "properties": {
                                    "price": {
                                        "type": "integer",
                                        "example": "20"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "About actions"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/actions/{id}/orders": {
            "get": {
                "tags": [
                    "Actions"
                ],
                "summary": "Display all action orders",
                "operationId": "actionsOrders",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of action",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ActionShowRequest"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/actions/{id}/products": {
            "get": {
                "tags": [
                    "Actions"
                ],
                "summary": "Display all action products",
                "operationId": "actionsProducts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of action",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ActionShowRequest"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/actions/{id}": {
            "get": {
                "tags": [
                    "Actions"
                ],
                "summary": "Get action by ID",
                "operationId": "actionsGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of action",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActionShowRequest"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Actions"
                ],
                "summary": "Update action by ID",
                "operationId": "actionsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of actions",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActionStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActionShowRequest"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Actions"
                ],
                "summary": "Delete action by ID",
                "operationId": "actionsDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of action",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ActionShowRequest": {
                "title": "Action get example",
                "description": "Some simple action example",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Unique ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "price": {
                        "title": "Price",
                        "description": "Value of price for storring",
                        "type": "integer",
                        "example": "60"
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title for storring",
                        "type": "string",
                        "example": "About action"
                    }
                },
                "type": "object"
            },
            "ActionStoreRequest": {
                "title": "Action post example",
                "description": "Some simple request createad as action",
                "properties": {
                    "price": {
                        "title": "price",
                        "description": "Price of action for storring",
                        "type": "integer",
                        "example": "55"
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title for storring",
                        "type": "string",
                        "example": "About action"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}